Move ANONYMOUS_AUTH to the constants file
