+++ src/components/AuthForm.js
@@ -5,8 +5,8 @@ import React, { PureComponent } from "react";
 
 import BaseForm from "./BaseForm";
 import { debounce, getServerByPriority, omit } from "../utils";
+import { ANONYMOUS_AUTH } from "../constants";
 
-export const ANONYMOUS_AUTH = "anonymous";
 const anonymousAuthData = server => ({
   authType: ANONYMOUS_AUTH,
   server: server,
