+++ src/constants.js
@@ -1,5 +1,6 @@
 /* @flow */
 export const DEFAULT_KINTO_SERVER = "https://kinto.dev.mozaws.net/v1/";
+export const ANONYMOUS_AUTH = "anonymous";
 
 export const ATTACHMENT_DELETE_REQUEST = "ATTACHMENT_DELETE_REQUEST";
